Extract calculateBorderTimings out
