+++ v3/src/js/types/timetable.js
